+++ src/main/java/duke/FileReaderHandler.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.exception.DukeException;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Todo;
