+++ src/main/java/duke/Storage.java
@@ -1,3 +1,5 @@
+package duke;
+
 import duke.task.Task;
 
 import java.io.File;
