+++ src/main/java/duke/UI.java
@@ -1,3 +1,5 @@
+package duke;
+
 import duke.task.Task;
 
 import java.io.File;
