+++ src/main/java/duke/command/CreateEventCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.UI;
+import duke.exception.DukeException;
 import duke.task.Event;
 
 import java.text.ParseException;
