+++ src/main/java/duke/exception/EmptyTodoTextException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class EmptyTodoTextException extends DukeException {
+    public EmptyTodoTextException(String message) {
+        super(message);
+    }
+}
