+++ src/main/java/duke/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
