+++ src/main/java/duke/exception/TaskDoesNotExistException.java
@@ -1,0 +1,8 @@
+package duke.exception;
+
+public class TaskDoesNotExistException extends DukeException {
+
+    public TaskDoesNotExistException(String message) {
+        super(message);
+    }
+}
